requirements.txt
# GCP Streaming & Event-Driven Telemetry
google-cloud-pubsub>=2.18.0
google-cloud-storage>=2.10.0

# PostgreSQL for Transformer Service
psycopg2-binary>=2.9.9
